requirements.txt
# App web
streamlit==1.44.1
streamlit-option-menu           # (optionnel mais utile pour l'UI)

# Traitement texte & data
numpy==2.2.5
pandas==2.2.3

# Manipulation audio/vidéo
ffmpeg-python==0.2.0
moviepy==1.0.3
av                              # pour la gestion vidéo de Streamlit

# YAML config 
pyyaml==6.0.1

# NLP & modèles 
transformers

# Utils
tqdm
joblib

# Transcription 
torch
torchvision
torchaudio
openai-whisper
ffmpeg-python
srt==3.5.3

#Traduction
googletrans==4.0.0rc1

# You also have to install the openai whisper model
# You can to it with this command line
# pip install git+https://github.com/openai/whisper.git --no-build-isolation
